video: a list is walked by the focus, not by the keyboard
Up and down in the game list moved its own selection from inside the
keyboard's handling of the page, so a pad standing on the same list
had nothing spend them on: it walked the map instead, found the rows
were all one box, and did nothing at all. The host-disk list, whose
scrolling was already the focus's, worked from either hand.

That step is now the focus's too, next to the host-disk one, and the
two hands mean the same thing by it. Home and End stay the keyboard's,
having no direction a pad could give them.
